Model ResourceInfoLayers for the resourceinfo slice and hide ResourceInfoRepo from all other slices
